refactor: move useTabPanel to its own folder.
