+++ packages/mui-base/src/TabPanelUnstyled/index.ts
@@ -3,5 +3,3 @@ export * from './TabPanelUnstyled.types';
 
 export { default as tabPanelUnstyledClasses } from './tabPanelUnstyledClasses';
 export * from './tabPanelUnstyledClasses';
-
-export { default as useTabPanel } from './useTabPanel';
