+++ packages/mui-base/src/index.js
@@ -20,6 +20,8 @@ export * from './useSlider';
 
 export * from './useSwitch';
 
+export * from './useTabPanel';
+
 export { default as BadgeUnstyled } from './BadgeUnstyled';
 export * from './BadgeUnstyled';
