+++ packages/mui-base/src/useTabPanel/index.ts
@@ -1,0 +1,5 @@
+export { default as useTabPanel } from './useTabPanel';
+export { default } from './useTabPanel';
+export * from './useTabPanel';
+
+export * from './useTabPanel.types';
